nessie-api-client/pyproject.toml
[tool.poetry]
name = "nessie-api-client"
version = "0.73.1-SNAPSHOT"
description = "A client library for accessing Nessie API"

authors = []

readme = "README.md"
packages = [
    {include = "nessie_api_client"},
]
include = ["CHANGELOG.md", "nessie_api_client/py.typed"]

[tool.poetry.dependencies]
python = "^3.8"
httpx = ">=0.20.0,<0.26.0"
attrs = ">=21.3.0"
python-dateutil = "^2.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target_version = ['py38', 'py39', 'py310', 'py311']
exclude = '''
(
  /(
    | \.git
    | \.venv
    | \.mypy_cache
  )/
)
'''

[tool.isort]
line_length = 120
profile = "black"
